Задача 2. Это и предыдущее решения неверные (эксперимент с целочисленными типами не удался)

+++ src/problem-002-add-two-numbers/ListNode.cs
@@ -19,7 +19,7 @@ public class ListNode
 
         var other = (ListNode)obj;
 
-        return this.ToInt() == other.ToInt();
+        return this.ToLong() == other.ToLong();
     }
 
     public override int GetHashCode()
